+++ CHANGELOG.md
@@ -5,6 +5,8 @@
 ## <next>
 * Don't allow sorting when grid is busy
 * Add scrollbar to dropdown menu
+* Add minWidth and maxWidth props to columns
+* Add default minWidth to prevent column to disappear
 * Fix bug: Default values are not applied to new items if grid is empty
 
 ## 2.1.3