+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 ## <next>
 * Align numbers and dates to right by default
+* Focus to new cell when adding new item
 * Don't allow sorting when grid is busy
 * Add scrollbar to dropdown menu
 * Add minWidth and maxWidth props to columns