+++ CHANGELOG.md
@@ -3,6 +3,8 @@
 * In general follow (https://docs.npmjs.com/getting-started/semantic-versioning) versioning.
 
 ## <next>
+
+## 2.2.0
 * Align numbers and dates to right by default
 * Don't allow sorting when grid is busy
 * Add scrollbar to dropdown menu